+++ NEWS.md
@@ -17,6 +17,7 @@
 * Added snapshot tests for plot methods of ERAHUMED layers (#149).
 * Added basic server tests for all shiny modules (#150).
 * Added basic UI no-error tests for all shiny modules.
+* Testing of `run_simulation()`.
 * Testing corner cases of assertions.
 * Testing of `print()` and `summary()` S3 methods for `erahumed_simulation` and 
 `erahumed_simulation_layer` S3 classes.